Build image using git_ref or use pre-built image (#3)
This PR contains config to either use pre-built image(s) or build
image(s) using the code specified by git_ref. This is controlled by
`global.build.enabled` property that is defined in `values-global.yaml`.
Current value of this property is set to `false` that configures the app
to use pre-built images. To build images from source set this property
to `true`